+++ cli/tests/test_immutable_commits.rs
@@ -579,4 +579,18 @@ fn test_rewrite_immutable_commands() {
     [EOF]
     [exit status: 1]
     "#);
+    // unsign
+    let output = test_env.run_jj_in(&repo_path, &["unsign", "-r=main"]);
+    insta::assert_snapshot!(output, @r#"
+    ------- stderr -------
+    Error: Commit bcab555fc80e is immutable
+    Hint: Could not modify commit: mzvwutvl bcab555f main | (conflict) merge
+    Hint: Immutable commits are used to protect shared history.
+    Hint: For more information, see:
+          - https://jj-vcs.github.io/jj/latest/config/#set-of-immutable-commits
+          - `jj help -k config`, "Set of immutable commits"
+    Hint: This operation would rewrite 1 immutable commits.
+    [EOF]
+    [exit status: 1]
+    "#);
 }